+++ fluent_contents/plugins/sharedcontent/models.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from future.builtins import str
 from future.utils import python_2_unicode_compatible
 from django.contrib.sites.models import Site
@@ -10,7 +9,7 @@ from .managers import SharedContentManager
 
 
 def _get_current_site():
-    return Site.objects.get_current()
+    return Site.objects.get_current().pk
 
 
 @python_2_unicode_compatible
@@ -22,7 +21,7 @@ class SharedContent(TranslatableModel):
         title = models.CharField(_("Title"), max_length=200)
     )
 
-    parent_site = models.ForeignKey(Site, editable=False, default=int(settings.SITE_ID))
+    parent_site = models.ForeignKey(Site, editable=False, default=_get_current_site)
     slug = models.SlugField(_("Template code"), help_text=_("This unique name can be used refer to this content in in templates."))
     is_cross_site = models.BooleanField(_("Share between all sites"), blank=True, default=False,
         help_text=_("This allows contents can be shared between multiple sites in this project.<br>\n"